+++ monitor/config_test.go
@@ -28,3 +28,25 @@ store-interval="10m"
 		t.Fatalf("unexpected store-interval:  %s", c.StoreInterval)
 	}
 }
+
+func TestConfig_Validate(t *testing.T) {
+	// NewConfig must validate correctly.
+	c := monitor.NewConfig()
+	if err := c.Validate(); err != nil {
+		t.Fatalf("unexpected validation error: %s", err)
+	}
+
+	// Non-positive duration is invalid.
+	c = monitor.NewConfig()
+	c.StoreInterval *= 0
+	if err := c.Validate(); err == nil {
+		t.Fatalf("unexpected successful validation for %#v", c)
+	}
+
+	// Empty database is invalid.
+	c = monitor.NewConfig()
+	c.StoreDatabase = ""
+	if err := c.Validate(); err == nil {
+		t.Fatalf("unexpected successful validation for %#v", c)
+	}
+}