+++ stress/v2/stress_client/commune_test.go
@@ -15,8 +15,12 @@ func TestCommunePoint(t *testing.T) {
 	if point.Tags().GetString("tag") != "tagVal" {
 		t.Errorf("expected: tagVal\ngot: %v", point.Tags().GetString("tag"))
 	}
-	if int(point.Fields()["fooField"].(float64)) != 5 {
-		t.Errorf("expected: 5\ngot: %v\n", point.Fields()["fooField"])
+	fields, err := point.Fields()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if int(fields["fooField"].(float64)) != 5 {
+		t.Errorf("expected: 5\ngot: %v\n", fields["fooField"])
 	}
 	// Make sure commune returns the prev point
 	comm.ch <- ""
@@ -27,8 +31,8 @@ func TestCommunePoint(t *testing.T) {
 	if point.Tags().GetString("tag") != "tagVal" {
 		t.Errorf("expected: tagVal\ngot: %v", point.Tags().GetString("tag"))
 	}
-	if int(point.Fields()["fooField"].(float64)) != 5 {
-		t.Errorf("expected: 5\ngot: %v\n", point.Fields()["fooField"])
+	if int(fields["fooField"].(float64)) != 5 {
+		t.Errorf("expected: 5\ngot: %v\n", fields["fooField"])
 	}
 }
 
@@ -43,7 +47,11 @@ func TestSetCommune(t *testing.T) {
 	if pt.Tags().GetString("tag") != "tagVal" {
 		t.Errorf("expected: tagVal\ngot: %v", pt.Tags().GetString("tag"))
 	}
-	if int(pt.Fields()["fooField"].(float64)) != 5 {
-		t.Errorf("expected: 5\ngot: %v\n", pt.Fields()["fooField"])
+	fields, err := pt.Fields()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if int(fields["fooField"].(float64)) != 5 {
+		t.Errorf("expected: 5\ngot: %v\n", fields["fooField"])
 	}
 }