+++ stress/v2/stress_client/reporting_test.go
@@ -57,7 +57,11 @@ func TestWritePoint(t *testing.T) {
 	if statementID != got {
 		t.Errorf("expected: %v\ngot: %v\n", statementID, got)
 	}
-	got2 := int(pt.Fields()["status_code"].(int64))
+	fields, err := pt.Fields()
+	if err != nil {
+		t.Fatal(err)
+	}
+	got2 := int(fields["status_code"].(int64))
 	if responseCode != got2 {
 		t.Errorf("expected: %v\ngot: %v\n", responseCode, got2)
 	}
@@ -80,7 +84,11 @@ func TestQueryPoint(t *testing.T) {
 	if statementID != got {
 		t.Errorf("expected: %v\ngot: %v\n", statementID, got)
 	}
-	got2 := int(pt.Fields()["status_code"].(int64))
+	fields, err := pt.Fields()
+	if err != nil {
+		t.Fatal(err)
+	}
+	got2 := int(fields["status_code"].(int64))
 	if responseCode != got2 {
 		t.Errorf("expected: %v\ngot: %v\n", responseCode, got2)
 	}